requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
yt-dlp==2023.11.16
pydantic==2.5.2
ratelimit==2.2.1
python-dotenv==1.0.0